(#22) Access to store `backend` is now entirely async
Moved all `backend` access to the `StoreHelpers`. Additionally the
backend access is now wrapped in promises, so localStorage and
asynchronous storage access can be handled via swapping out the backend,
rather than overriding methods on Store or MetaStore.

Adding configuration to the Integration tests to allow running them with
different backends. This should help when testing backend adapters for #6.